refactored (not completed) added market, custom conditions , fixed tests

+++ priv/repo/migrations/20230327124704_create_bookings.exs
@@ -8,7 +8,7 @@ defmodule Ukio.Repo.Migrations.CreateBookings do
       add :apartment_id, references(:apartments, on_delete: :delete_all), null: false
       add :monthly_rent, :integer
       add :deposit, :integer
-      add :utilities, :integer
+      add :utilities, :float
 
       timestamps()
     end
